api/requirements.txt
requests
mysql-connector-python
dateparser
pytest-playwright
websocket-server
python-dotenv